+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(meta)
+
+test_check("meta")